infra/assets/extra-requirements.txt
# Azure Functions core
azure-Functions>=1.24.0

# GitHub Copilot SDK
github-copilot-sdk==0.1.29

# Pydantic for tool parameter validation
pydantic>=2.0.0

# Parse AGENTS.md frontmatter and body
python-frontmatter>=1.1.0

azurefunctions-extensions-http-fastapi

# Azure Identity for Entra ID (managed identity) authentication
azure-identity>=1.17.0

# OpenAI SDK for direct Azure OpenAI calls (bypassing Copilot SDK)
openai>=1.40.0

# Bot Framework SDK for Microsoft Teams integration
botbuilder-core>=4.16.1
botbuilder-integration-aiohttp>=4.16.1
botframework-connector>=4.16.1

# Azure Blob Storage for session state persistence (serverless)
azure-storage-blob>=12.19.0

# Async HTTP client for Speech REST API calls and Teams attachment downloads
aiohttp>=3.9.0

# Microsoft Authentication Library for Graph OBO flow
msal>=1.30.0